Add extra file for tests
